ejer7

+++ Taller10/src/paquete8/Problema08.java
@@ -13,15 +13,15 @@ public class Problema08 {
     public static void main(String[] args) {
         // Creación de arreglo bidimensionales
 
-        double[][] dato1 = {{1, 2, 3}, {6, 8, 9}};
-        double[][] dato2 = {{10, 1, 2}, {10, 9, 1}};
+        int[][] dato1 = {{1, 2, 3}, {6, 8, 9}};
+        int[][] dato2 = {{10, 1, 2}, {10, 9, 1}};
         // En función de los arreglos dados, genere
         // un arreglo que tenga los siguientes valores
         /*
             1   1   2
             6   8   1
          */
-        double[][] resultado = new double[2][3];
+        int[][] resultado = new int[2][3];
 
         for (int i = 0; i < resultado.length; i++) {
             for (int j = 0; j < resultado[i].length; j++) {
@@ -35,11 +35,10 @@ public class Problema08 {
         }
 
         System.out.println("Resultado final:");
-        
-        
+
         for (int i = 0; i < resultado.length; i++) {
             for (int j = 0; j < resultado[i].length; j++) {
-                System.out.printf("Posicion[%d][%d] con valor de: %.2f\n", 
+                System.out.printf("Posicion[%d][%d] con valor de: %d\n",
                         i,
                         j,
                         resultado[i][j]);
